Errata source UX fixes
Issue #899
Set width to 100%
Change label
Change checkbox to yes/no radios
Add expand/collapse indicators to options
Hide 2nd button when both are empty
